Introduce Repository class
In many scenarios for testing scripts that are meant to be used in
conjunction with git it is useful to test them on a git repository
directly.
To that end, this change introduces a new class, 'Repository', that
wraps a git repository useful for such test cases. In particular, the
class takes care to create a new and empty directory, to initialize the
repository correctly, and to provide common repository related actions
(such as adding of a file or creating a commit), while avoiding
interference by any custom user configuration (think, custom hooks, GPG
signing of commits, user name and e-mail, and more) present on the
system.