Only change into repository if not already in it
For testing the behavior of paths specified in a relative manner for
various git extensions, it is useful to be able to invoke a command from
a sub-directory of a repository instead of always the root.
This change detects whether we are currently already inside the
repository and only if not changes the current working directory into
the repository's root. Otherwise it is left untouched.