Use execute module for process execution
Until now we used the subprocess module for starting new processes (in
order to run git commands). We want to move away from this module and
instead use the execute module. The latter module provides a more
user-friendly interface and in the future it will support starting
processes with a custom environment, which helps in making the
'Repository' class more safe to use in conjunction with other process
invocations.
To that end, this change converts the source code to make use of the
execute module.