Use custom environment when invoking git command
With the previously used subprocess module the invoking process'
environment is unconditionally inherited to the child. Among others,
this inheritance causes the HOME variable to be defined in the child
process. Because in our context we want to achieve isolation from the
invoking user's configuration we specifically unset HOME.
With recent changes we switched to using the execute module and this
module's process execution functionality supports spawning a process
with a custom environment.
To that end, this change removes the logic for unsetting the HOME
environment variable and, instead, simply uses an empty environment for
the newly spawned process. Using this approach we get better isolation
from any user-defined configuration with fewer lines of code.