Add mixins to ease inheriting environment variables to child processes
When running git in an empty environment we can potentially run into
problems if the git invocation causes other programs to run. In
particular, if a git invocation causes a hook written in Python to run
we might need a PYTHONPATH variable to be available (at least if the
hook has non-standard dependencies it needs to import).
This change adds mixins to support such scenarios. In particular, it
adds the PathMixin and the PythonMixin classes that cause commands
executed on behalf of the repository to inherit the PATH and PYTHON*
environment variables, respectively.
Note that the PYTHON* variables' contents are taken directly from the
parent process' environment, not from the representations inside a
Python instance (that is, for example sys.path for PYTHONPATH and
sys.dont_write_bytecode for PYTHONDONTWRITEBYTECODE). The reason is that
the Python instance's contents might have been modified in a way not
intended/anticipated/wanted by the original caller. We always assume the
caller to have the "last word" in those matters.